backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
langchain==0.1.11
langchain-anthropic==0.1.4
aiosqlite==0.20.0
python-dotenv==1.0.1
httpx==0.27.0
python-multipart==0.0.9
python-json-logger==2.0.7
